Safari-609.1.11.1/LayoutTests/fast/dom/HTMLDocument/set-focus-on-valid-element-expected.txt

document.activeElement should return a valid and focusable element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.activeElement is not willBeDisabled
PASS document.activeElement.id is "body"
PASS successfullyParsed is true

TEST COMPLETE
